Fix scoring ICA assessments
OppId input parameter was being ignored and a value for OppId was being generated instead when adding a row to XMLRepository database.
This is due to TDS sending a zero value for the OppId in the TRT.  TIS was expecting a value OppId value.
This change detects if a zero value for OppId is being sent and generate a valid OppId.
Otherwise, the OppId being passed in is used.
This is important for ICA assessments, since a new, virtual exam is created and the OppId created and set in a different table (CombinationTestOpportunity).