fix changes in scale score from existing test specification to new format
The scale score is computed by multipling the theta score to the item selection slope and adding the intercept value.
When loading ICA documents the slope and intercept values were set to 1 for both values.

The new format does not have an item selection section for the test package level (only at the test level).
Using the first slope and intercept values found in a test.