fixed error on TIS windows service startup
When ICA assessments with multiple segments are loaded in the new assessment format, the TIS windows service does not start.
The error message given:  "The GPC model must have one more parameter than score points but this item has scorePoints = 2 and 5 parameters"
This is due to too many rows added to the ItemMeasurementParameter table when an item has a dimension.
Adding filtering to the number item measurement parameters based on the parameter dimension field.